Added Draken Creation Workshop Host config files
The Draken has 8 projector mounting positions, each coorelating to
a different X,Y resolution setting.  So this essentially results in
8 different "printers" in the CWH configuration system.

Below are the 8 different resolutions currently configured, mapped to
the projector mounting slot, where 1 == top most, and 8 == bottom most.

Postion:        XY Resolution:
   1                 37µm
   2                 43µm
   3                 50µm
   4                 61µm
   5                 70µm
   6                 75µm
   7                 82µm
   8                 90µm

Currently, all the *.slicing files are identical, as there is a bug in
CWH that will only open a .slicing file that matches the .machine name
specified in the .printer file.

When this bug is fixed (CWH issue #46: https://goo.gl/CNhZd4), then we
should be able to revert to using a single .slicing file.